### 2016-07-12 #38 @whedon
[REVIEW]: DoSOCS: A System for Managing Open Source Risk
**Submitting author:** @germonprez (<a href="http://orcid.org/0000-0003-2326-5901">Matt Germonprez</a>)
**Repository:** <a href="https://github.com/DoSOCSv2/DoSOCSv2" target ="_blank">https://github.com/DoSOCSv2/DoSOCSv2</a>
**Version:** 0.16.1
**Archive:** <a href="https://dx.doi.org/10.6084/m9.figshare.4239665.v1" target="_blank">https://dx.doi.org/10.6084/m9.figshare.4239665.v1</a>
