### 2016-07-12 #38 @arfon
Thanks for submitting this @germonprez! It looks like this submission is missing the JOSS paper (`paper.md`). Could you take a look at our [submission guidelines](http://joss.theoj.org/about#author_guidelines) on how to structure the paper and let me know when you've added it to your repo?
