### 2016-10-10 #38 @germonprez
@arfon I've uploaded the paper.md paper.bib, and codemeta.json in the paper folder:

https://github.com/DoSOCSv2/DoSOCSv2/tree/master/paper

Happy to make changes as needed. Thanks!
