### 2016-10-10 #38 @mlinksva
:hand: I'll review this, thanks for ping @arfon.

I haven't investigated whether the bot only responds to certain users, but I'll try anyway:
