### 2016-10-10 #38 @arfon
> ✋ I'll review this, thanks for ping @arfon.

Wonderful, many thanks @mlinksva.

> I haven't investigated whether the bot only responds to certain users, but I'll try anyway:

Nice try @mlinksva. @whedon should have given you an error but it looks like he failed silently.
