### 2016-10-12 #38 @mlinksva
> Version: Does the release version given match the GitHub release (v2.0)?

I'm not sure. It seems that "v2" here is not really a software version, but part of the name of the software, indicating that there was a previous implementation -- the v2 suffix plays the same role a neo prefix would. The current version [seems to be](https://github.com/DoSOCSv2/DoSOCSv2/blob/62f48ffbea96e7befe113ea0298691d2ccfdc96c/setup.py#L3) 0.16.1. Also the README says "dosocs2 is under heavy development; expect frequent backwards-incompatible changes until a 1.x.x release!"

@germonprez should the version submitted be 0.16.1?
