### 2016-10-12 #38 @germonprez
Yes, mistake on my end. The v2 in the name is the reboot of DoSOCS. We had an original version v1 that was really proof of concept. The v2 is the second iteration on DoSOCS -- a fundamental rewrite.

You read this correctly.
