### 2016-10-12 #38 @mlinksva
Not really pertinent to review, but for my future reference, example of an Apache-2.0->GPL-3.0+ license change with rationale: https://github.com/DoSOCSv2/DoSOCSv2/commit/8704053fb10d3238751bba95302c4c343e0c0404#diff-9879d6db96fd29134fc802214163b95a
