### 2016-10-12 #38 @mlinksva
Great, I changed the version field above to 0.16.1.
