### 2016-10-18 d16bc29
Fix the error as the following:
File
"/home/leimh/.local/lib/python2.7/site-packages/sqlalchemy/engine/default.py",
line 447, in do_executemany
    cursor.executemany(statement, parameters)
sqlalchemy.exc.ProgrammingError: (sqlite3.ProgrammingError) You must not
use 8-bit bytestrings unless you use a text_factory that can interpret
8-bit bytestrings

Signed-off-by: zhengruoqin <zhengrq.fnst@cn.fujitsu.com>
