### 2016-11-09 #38 @mlinksva
Quick comments on the paper before I get to the software itself.

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Could be stronger. States what software does from which problem can be inferred, target audience is "in an organization" with those problems. Skimming the first page of accepted papers at http://joss.theoj.org/papers/popular it seems most provide a bit more detail on the problem/need addressed and/or how submitted software relates to other software.

Also, the paper includes a reference to the software the paper is about. It doesn't look like any accepted papers have this.
