### 2016-11-09 #38 @mlinksva
Under documentation:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

https://github.com/DoSOCSv2/DoSOCSv2/blob/62f48ffbea96e7befe113ea0298691d2ccfdc96c/CONTRIBUTING.md says "The contribution rules for DoSOCS are currently being rewritten. You are welcome to create issues, but we can't accept pull requests from the community at this time." I suppose this technically does all of 1-3 if "can't accept" works for 1 and "welcome to create issues" also implies this is how to seek support. @arfon I could use some guidance on whether this is satisfactory for checking off community guidelines.
