### 2016-11-10 #38 @arfon
> @arfon I could use some guidance on whether this is satisfactory for checking off community guidelines.

This isn't great but I think it _does_ meet the requirements for 1--3.
