### 2016-11-14 #38 @mlinksva
Great, I see in https://github.com/DoSOCSv2/DoSOCSv2/pull/94

I don't know how the paper PDF is generated, but I've gone ahead and checked off the relevant items.
