### 2016-11-14 #38 @mlinksva
Remaining checklist items.

**Functionality: Have the functional claims of the software been confirmed?**

README claims:

> dosocs2 is a command-line tool for managing SPDX 2.0 documents and data. **It can scan source code distributions to produce SPDX information, store that information in a relational database, and extract it in a plain-text format on request.**
>
> The discovery and presention of software package license information is a complex problem facing organizations that rely on open source software within their innovation streams. **dosocs2 enables easy creation of a "bill of materials" for any software package to represent associated license information.** In addition, dosocs2 and can be used in the creation and continuous maintenance of an inventory of all open-source software used in an organization.

Scanning/storing/extracting it can plainly do (though one has to remember/guess/look in database without help from dosocs2 for integer package id in order to do the last, unless I'm missing something).

I'm not sure about easy creation of a bill of materials, unless raw SPDX documents are taken to be a bill of materials. I suspect this may just be a documentation issue: state that SPDX documents are intended to serve as BOM, or the easy part is that the info is in a database that one can query with SQL to generate whatever BOM or other report one wants. An end-to-end example of scanning, BOM generation, and continuous maintenance would be really helpful. Perhaps the dosocs2 itself could be the subject of the end-to-end example.

It would also be helpful to document how an organization would use dosocs2 in relation to other tools, or not. dosocs2 includes a component from fossology. Would an organization using fossology have any reason to also use dosocs2? How?

**Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).**

See above.

**Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g. API method documentation)?**

`--help` outputs acceptable manpage style documentation.

`/doc` contains some useful info on what the available scanners do.

If user needs to query database for reporting (see above), would be nice to have that schema documented.

**Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?**

It does have tests in `tests` which are run by CI. Don't know how comprehensive. Ideally running and adding tests might be documented in `CONTRIUBTING.md`. I'm going to guess that what's already there is satisfactory for this review checkoff.
