### 2016-11-15 #38 @germonprez
Updates have been made and merged to the README.

1) removed the 'easy BOM' statement and just declared that it produces an SPDX document
2) included a DFD to represent a potential organizational use of dosocs2. It is quite possible that a company would use FOSSology and dosocs2. FOSSology is really a 1-off check of a software package. Dosocs2 is built with the intention of coordinating discovery as part of a larger organizational workflow.

I didn't really understand the suggestion: Perhaps the dosocs2 itself could be the subject of the end-to-end example.
