### 2016-11-15 #38 @mlinksva
Thanks. Suggestion was showing how dosocs2 could "be used in the creation and continuous maintenance of an inventory  of all open-source software used in" itself.

In any case I'm fine with checking off the remaining items now. Not having done a review here before, @arfon can I get a second opinion?
