+++ README.md
@@ -11,10 +11,10 @@ scan source code distributions to produce SPDX information, store that
 information in a relational database, and extract it in a plain-text format
 on request.
 
-The discovery and presention of software package license information is a complex
+The discovery and presentation of software package license information is a complex
 problem facing organizations that rely on open source software within their 
 innovation streams. dosocs2 enables creation of an SPDX document for any 
-software package to represent associated license information. In addition, dosocs2 and 
+software package to represent associated license information. In addition, dosocs2 
 can be used in the creation and continuous maintenance of an inventory of all 
 open-source software used in an organization. The primary audience for dosocs2 is open source
 software teams seeking to advance the representation and maintenance of open source 
