### 2016-11-15 #38 @mlinksva
Oh, one other thing, the title "DoSOCS: A System for Managing Open Source Risk" seems an exaggeration -- the DFD shows it is a tool that could be part of a system for managing open source risk. Also missing "v2". "DoSOCSv2: A System for SPDX 2.0 Document Creation and Storage" would seem more accurate to me.

Finally, just out of curiosity, what does DoSOCS stand for/expand to? Just realized I don't know.
