### 2016-11-15 bb57f50
Update paper.md
