### 2016-11-15 91e5d87
Update paper.md
