### 2016-11-15 #38 @germonprez
Ok. These are fixed and being merged.

Regarding the name:

It used to be called DoSPDX when it was first a contribution to the Yocto project
We moved off the SPDX for a few reasons
We then went to SOCS (simple open compliance system) and kept the Do part

the v2 part is because it was a full rewrite of an earlier version of DoSOCS that had many technical problems. It just sort of stuck from that point!
