### 2016-11-15 #38 @mlinksva
Thanks @germonprez. Not important for this review, but I'd suggest adding that background trivia the project's README#History ... makes the project a bit more approachable IMO.
