### 2016-11-18 #38 @arfon
> In any case I'm fine with checking off the remaining items now. Not having done a review here before, @arfon can I get a second opinion?

@mlinksva - yep, I think this is looking good. Thanks for checking (and reviewing 😄 )
