### 2016-11-18 #38 @germonprez
Thanks @arfon and @mlinksva :smiley:

@arfon does this work: https://dx.doi.org/10.6084/m9.figshare.4239665.v1
