### 2016-11-18 #38 @arfon
@mlinksva many thanks again for the thorough review.

@germonprez - your paper is now accepted. Your DOI is http://dx.doi.org/10.21105/joss.00038 🚀 🎉 💥
