### 2016-11-18 #38 @arfon
@germonprez - one issue I didn't catch during the processing is that you have my ORCID listed in the `paper.md` (https://raw.githubusercontent.com/DoSOCSv2/DoSOCSv2/master/paper/paper.md). If the authors have an ORCIDs could you please update the `paper.md` file with these? If not, please update and remove my ORCID so it's not associated with you 😁
